+++ project/src/main/java/game/model/entity/projectile/Projectile.java
@@ -7,18 +7,18 @@ import javafx.geometry.Point2D;
 import javafx.scene.effect.Light;
 
 public abstract class Projectile<T extends IShape2D> extends MovableEntity<T> implements IProjectile<T> {
-    private final double strength;
+    private final int strength;
     private boolean isDestroyed;
 
 
-    public Projectile(Point2D position, double maxForce, double maxSpeed, double strength, Point2D velocity, T shape) {
+    public Projectile(Point2D position, double maxForce, double maxSpeed, int strength, Point2D velocity, T shape) {
         super(position, velocity, maxForce, maxSpeed, shape);
         this.strength = strength;
         this.isDestroyed = false;
     }
 
     @Override
-    public double getStrength() {
+    public int getStrength() {
         return strength;
     }
